Implement sync in PeerConnection

+++ internal/rpc/remote_rpc.go
@@ -12,5 +12,5 @@ type RemoteRPC interface {
 	GetChainID(ctx context.Context) (id *multihash.Multihash, err error)
 	GetHeadBlock(ctx context.Context) (id *multihash.Multihash, height uint64, err error)
 	GetAncestorBlockID(ctx context.Context, parentID *multihash.Multihash, childHeight uint64) (id *multihash.Multihash, err error)
-	GetBlocks(ctx context.Context, headBlockID *multihash.Multihash, startBlockHeight uint64, batchSize uint64) (blocks []protocol.Block, err error)
+	GetBlocks(ctx context.Context, headBlockID *multihash.Multihash, startBlockHeight uint64, batchSize uint32) (blocks []protocol.Block, err error)
 }
